Optamized Property List as well as fix Property price bug
