+++ airbnb/src/component/QueryRequirment/QueryRequirment.css
@@ -1,0 +1,4 @@
+.query-requirment {
+        margin-top: 50px;
+        margin-bottom: 50px;
+}
