+++ airbnb/src/component/QueryRequirment/QueryRequirment.js
@@ -1,0 +1,19 @@
+import React, { useState } from 'react';
+import './QueryRequirment.css';
+
+const QueryRequirment = () => {
+        return (
+                <div className='main query-requirment'>
+                        <div>
+                                <h2>Query Requirments</h2>
+                                <p>
+                                        This page is dedicated to all the query
+                                        requirments that didn't fit with our
+                                        Applicaiton.
+                                </p>
+                        </div>
+                </div>
+        );
+};
+
+export default QueryRequirment;
